Control board via HTML
- Serve the game via server
- GET /board - returns a board state in json
- POST /move - moves a piece from one square to another(json)
- Any other path that's GET is interpreted as
a path to a local file.

UI part is inside `ui` directory where `index.html`
is present with some styling and javascript.
Javascript allows drag and dropping the pieces,
those are sent to the server and new state is fetched
and rendered.

No rules, no validation, just a simple scaffolding